# app/src/main/resources/openapi.yml
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "localhost"
        }
    ],
    "paths": {
        "/jobs": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get All Jobs",
                "operationId": "get_all_jobs_jobs_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Post Job",
                "operationId": "post_job_jobs_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_post_job_jobs_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/jobs/{jobId}": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get Specific Job",
                "operationId": "get_specific_job_jobs__jobId__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Jobid",
                            "type": "integer"
                        },
                        "name": "jobId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Job"
                ],
                "summary": "Update Specific Job",
                "operationId": "update_specific_job_jobs__jobId__put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_update_specific_job_jobs__jobId__put"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/jobs/{jobId}/proposals": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get Proposals",
                "operationId": "get_proposals_jobs__jobId__proposals_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Jobid",
                            "type": "integer"
                        },
                        "name": "jobId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Create Proposal",
                "operationId": "create_proposal_jobs__jobId__proposals_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_proposal_jobs__jobId__proposals_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/jobs/{jobId}/proposals/{proposalId}": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get Specific Proposal",
                "operationId": "get_specific_proposal_jobs__jobId__proposals__proposalId__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Jobid",
                            "type": "integer"
                        },
                        "name": "jobId",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Proposalid",
                            "type": "integer"
                        },
                        "name": "proposalId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Job"
                ],
                "summary": "Update Proposal",
                "operationId": "update_proposal_jobs__jobId__proposals__proposalId__put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_update_proposal_jobs__jobId__proposals__proposalId__put"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/jobs/{jobId}/proposals/process": {
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Process Proposal",
                "operationId": "process_proposal_jobs__jobId__proposals_process_post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Jobid",
                            "type": "integer"
                        },
                        "name": "jobId",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Proposalid",
                            "type": "integer"
                        },
                        "name": "proposalId",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Status",
                            "type": "integer"
                        },
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User",
                    "User"
                ],
                "summary": "Login",
                "operationId": "login_user_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_user_login_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "User",
                    "User"
                ],
                "summary": "Create User",
                "operationId": "create_user_user_create_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_user_user_create_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_create_proposal_jobs__jobId__proposals_post": {
                "title": "Body_create_proposal_jobs__jobId__proposals_post",
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string",
                        "default": ""
                    },
                    "body": {
                        "title": "Body",
                        "type": "string",
                        "default": ""
                    }
                }
            },
            "Body_create_user_user_create_post": {
                "title": "Body_create_user_user_create_post",
                "required": [
                    "fullname",
                    "username",
                    "password",
                    "timezone"
                ],
                "type": "object",
                "properties": {
                    "fullname": {
                        "title": "Fullname",
                        "type": "string"
                    },
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    },
                    "timezone": {
                        "title": "Timezone",
                        "type": "integer"
                    },
                    "role": {
                        "title": "Role",
                        "type": "integer",
                        "default": 2
                    }
                }
            },
            "Body_login_user_login_post": {
                "title": "Body_login_user_login_post",
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
            "Body_post_job_jobs_post": {
                "title": "Body_post_job_jobs_post",
                "required": [
                    "title",
                    "desc"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "desc": {
                        "title": "Desc",
                        "type": "string"
                    },
                    "published": {
                        "title": "Published",
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "Body_update_proposal_jobs__jobId__proposals__proposalId__put": {
                "title": "Body_update_proposal_jobs__jobId__proposals__proposalId__put",
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string",
                        "default": ""
                    },
                    "body": {
                        "title": "Body",
                        "type": "string",
                        "default": ""
                    }
                }
            },
            "Body_update_specific_job_jobs__jobId__put": {
                "title": "Body_update_specific_job_jobs__jobId__put",
                "required": [
                    "title",
                    "desc"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "desc": {
                        "title": "Desc",
                        "type": "string"
                    },
                    "published": {
                        "title": "Published",
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "JWTBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}